+++ backend/server/presentation/src/main/kotlin/com/gazim/gmessenger/server/utils/Utils.kt
@@ -3,5 +3,5 @@ package com.gazim.gmessenger.server.utils
 import io.ktor.server.application.*
 
 interface ISecurityUtils {
-    suspend fun getUserTokenId(call: ApplicationCall): Long
+    suspend fun getUserTokenId(call: ApplicationCall): Int
 }
